config: Use public package repo for buildroot
This allows the image builder to be used outside of the Endless VPN. As
a bonus it will almost certainly be faster except for builds running in
the same AWS region as OBS.

There are a couple caveats, though. The public repo filters out a few
non-redistributable packages. This should typically not be an issue as
these won't be used during the buildroot construction. In case they are,
the local build can override to use the internal repository. The other
issue is that there is some latency between when the packages are
published on the internal repo to when they're available in the public
repo. Currently this is a maximum of 10 minutes, which should be
acceptable for most use cases.

https://phabricator.endlessm.com/T31225