Remove the check_update stage
Back in ye olden days, all of the image builds ran serially on one
machine, so there was benefit to skipping builds when it didn't seem
like there was anything to do. Thus the check or check_update stage that
tries to look at the previous build, determine if anything will be
different from this build, and exit early if nothing appeared different.

Times have changed and this seems like more trouble than it's worth.

* It only checks a few things, so it will decide things haven't changed
  when they will. Things like what apps will be installed, for instance.

* The implementation is somewhat fragile. If there was an error in one
  of the check hooks, should the build abort? If not, would you ever
  notice if the checks weren't working correctly?

* Finding out details of the previous image build depends on our
  internal image server's API, which isn't going to be available
  publicly. We're trying to make this usable by the public, so that's
  not going to work.

* Nowadays we have a better build infrastructure with several builders
  that can be run in parallel. If a potentially unneeded build is done,
  I don't think anyone cares. And if someone manually starts a build,
  they probably really want that build to happen and not have the
  builder tell them it knows better.

This removes the stage, hooks and a couple other pieces that were there
only to support it.

https://phabricator.endlessm.com/T31274