Only allow one personality per build
Back in the day the ostree and image builds were done in the same tool
with the images built serially after the ostree. It made sense then to
build all the personalities for the specific product/arch/platform
combination in a loop.

Later the ostree builder was split out and the image builder always
pulls the ostree from the remote repository instead of using the one on
disk. In that case there's little benefit to building multiple images in
one run. Although the interface to loop over personalities was there,
our driver in jenkins has specified a single personality as long as I
can remember. Now that we're opening up the tool, it's time to remove
that historical artifact.

https://phabricator.endlessm.com/T31274