Remove checkout options
Let it always be the caller's responsibility to checkout the appropriate
branch of the builder. This behavior was clever in the early days to get
the builder to automatically use the right branch when it was running
from a fixed checkout, but nowadays our driver in jenkins always does
the checkout and has been using the `--no-checkout` option to ensure
this code never runs.

https://phabricator.endlessm.com/T31274