Remove the --use-production-apps option
This has actually been a no-op for a while because it would depend on
the flatpak remote configurations using a different stage for pulling
and deploying. But we actually stopped doing that quite a while ago
since the apps and SDK development process is not coupled to the OS
development process.

https://phabricator.endlessm.com/T31288