config: Remove authentication for ostree URLs
The OS repos no longer require authentication, so now all of the ostree
URLs in the image builder can be used without authentication. That setup
has been removed and the URLs have been reworked (again) so that
everything comes from the CDN by default with no credentials.

The ability to specify a separate pull URL has been retained. For our
builders that are in the same AWS region as the ostree server, they can
use our internal ostree.endlessm-sf.com domain and bypass the CDN for
faster access.

https://phabricator.endlessm.com/T31288