eib: Unmount root itself in unmount_root_filesystems
Previously we tried to skip unmounting the root path itself under the
idea that the image builder wasn't doing that and if it was a mount
point that maybe someone did that outside the builder. Now we want to
make the buildroot itself a mount and cleanup with this code, so drop
that precaution.

https://phabricator.endlessm.com/T31516