Don't delete output directory
Especially now that no publishing happens by default, the output files
shouldn't be deleted upon completion. This was important in the days
when multiple personalities were built from a single run so that the
disk wouldn't fill up. Only a single personality is built now and the
output directory is deleted at the beginning of the next build, so the
disk usage shouldn't affect subsequent builds.

https://phabricator.endlessm.com/T31698