flatpak: Use the default architecture from the buildroot flatpak
Originally the flatpak architecture was needed in a few spots, so having
it set in the configuration was convenient. Also, at that time, `flatpak
--default-arch` did not exist, IIRC. So, the CPU from
`dpkg-architecture` was used as an approximation of the flatpak
architecture to use.

Instead, leave the config setting available but empty by default and
have the only remaining user get the default architecture from flatpak,
which will be available in the buildroot. If necessary, a shell helper
that checks `$EIB_FLATPAK_ARCH` and falls back to `flatpak
--default-arch` can be added.

https://phabricator.endlessm.com/T31702