image: Fix race between ISO and checksum creation
In d8e43c39272d39131a560ff2f2be61ca556d77e3 the split image process was
moved after the ISO creation since it destructively edits the raw image.
However, this also moved the `wait` that was blocking on the
uncompressed signature file which the ISO creation needed. This
sometimes causes a race if the signature isn't completed before
`create-iso` is called.

Unfortunately, previously the race would be silently ignored because
`create-iso` assumes that if the signature file didn't exist it was
intentional. However, now it has a hard requirement on the uncompressed
image checksum file, which is also created in the background at the same
time.

This adds back the wait with a slight optimization that it only happens
when creating an ISO.

https://phabricator.endlessm.com/T31304