image: Drop merged flatpak repo setup
This was used to ensure that flatpak's repo at `/var/lib/flatpak/repo`
could be merged with `/ostree/repo`. We don't want that configuration
anymore, so it can be dropped.

With the repos disconnected, we no longer have to disable the OS remote
since the flatpak repo would never have it in the first place. Likewise,
no symlink juggling is needed to access the installation outside of the
chroot.

https://phabricator.endlessm.com/T31980