hooks: Remove 50-google-chrome and default to Chromium instead
Now we can ship `org.chromium.Chromium.Codecs` out of the box in many
cases, and in cases where we can't, it’ll be downloaded with app updates
relatively quickly when the machine goes online (similar to how Chrome
was silently installed when a network connection became available).

So we can use Chromium as the default browser everywhere, and have a
similar experience.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

https://phabricator.endlessm.com/T32407