Revert "Skip flatpak parental controls checks when running from the builder"
This reverts commit 2c1c0919dad298c538dcafae4b60720022177e68.

Due to T32545, which has landed on `master` and `eos4.0`, flatpak now
always ignores parental controls when run as the `root` user. This means
that we should no longer need this workaround for ignoring parental
controls in the image builder (which runs as `root`).

This commit can be backported to `eos4.0`, but no earlier. The version
of flatpak in earlier EOS branches does not have the fix for `root`.

https://phabricator.endlessm.com/T27896