kolibri: Clear local dir before downloading content
When the hook that downloads Kolibri content runs, its target directory
may already exist and have content from the previous run, which may be
from a different image configuration.

This commit removes any previously exiting content to prevent that
content from being included in the image being built.

Note that this defeats the purpose of the image builder saving the
contents of EIB_CONTENTDIR between runs to improve its performance.  We
should revisit this to try to make the caching work properly at some
point in the near future.

https://phabricator.endlessm.com/T32712