eos-image-builder: Don't change working directory
If the `--localdir` option is a relative path, then it will be
mishandled by `run-build` if the working directory is changed using the
`subprocess` `cwd` parameter.

Instead, build an absolute path to `run-build` and execute that. There's
no longer any need to change to the `eos-image-builder` directory since
the git checkout isn't changed.