ImageBuildRoot: Disable buildroot apt cache handling
The mmdebstrap hooks were doing entirely the wrong thing by syncing
between the image builder's apt cache and the **system's**
`/var/cache/apt/archives`. That's both wrong and pointless.

Changing that to sync into the buildroot's apt archives makes
`mmdebstrap` throw an error. Apparently it uses the contents of
`/var/cache/apt/archives` to figure out what to install after telling
apt to download packages, so it requires the directory to be empty when
it starts.

To bring this back, something like apt-cacher that works at the HTTP
proxy level is needed.