kolibri: Use our content CDN during builds
To offload our Learning Equality's CDN, we've setup our own CDN in front
of it at https://kolibri-content.endlessos.org. Import content using
that as the base URL. Installations are still configured to use the
Kolibri default URL. The configuration can be cleared to go back to the
default URL for builds, too.

https://phabricator.endlessm.com/T30950