ImageBuildRoot: Use ostree checkout for buildroot
Currently the buildroot is constructed from scratch using `mmdebstrap`.
That requires knowing the exact apt sources that the branch of the OS
uses. Instead, the buildroot can be created from a checkout of the
ostree that's going to be used for the deployment since it already
contains the appropriate apt/dpkg configuration.

That changes the host requirements from `mmdebstrap` and `rsync` to the
`ostree` CLI package. Ideally an `eos` container would be used rather
than the bespoke `chroot` setup, but that's a bigger change for another
day.

https://phabricator.endlessm.com/T35019