image: Specify bootloader to ostree
Until now we've relied on ostree's automatic bootloader detection, which
relies on the presence of certain files to determine which bootloader is
in use. Instead we should specify what bootloader to use so the decision
is unambiguous.

When grub is the bootloader, we actually don't want to use ostree's grub
integration. That tries to update grub.cfg whereas we just want it to
update the BLS entries since our grub can read those and our grub.cfg
doesn't support updating. This will help us drop an ugly downstream hack
that makes the ostree grub bootloader integration do nothing.

https://phabricator.endlessm.com/T33364