image: Make kolibri importcontent more robust
By default, `importcontent` ignores errors and has a 60 second timeout
for object requests. This is to make the tool more friendly for users on
unreliable networks. At runtime Kolibri will fetch any objects it's
missing.

In the image builder we want to ensure that all the imported channels
are fully provisioned. Similarly, we'd much rather wait a bit longer to
complete a request than have it fail early. This is particularly
important when using our own kolibri-content CDN since it blocks while
it fetches missing objects from Learning Equality's origin.

Kolibri 0.15.3 adds 2 `importcontent` options to address these. Add the
`--fail-on-error` option and use a 5 minute `--timeout`.

https://phabricator.endlessm.com/T33363
https://phabricator.endlessm.com/T33365