eibflatpak: Rewrite to use Flatpak transactions
The bespoke method of resolving and installing Flatpaks was difficult to
maintain, used deprecated Flatpak APIs, and ultimately became broken for
arm64. This rewrites the `FlatpakManager` class to use Flatpak
transactions. That takes care of all the dependency resolution and
installation ordering.

A couple tricks are still needed to preserve previous features. First, a
manual OSTree pull from the cache repo is still needed to seed the
commit objects in the image installation. Second, in order to skip
excluded related refs (i.e., extensions), the Flatpak mask feature is
used.

https://phabricator.endlessm.com/T33827