config: move Rhythmbox from mandatory to default
The 'mandatory' lists of apps are (a superset of) those which are
installed on existing systems when they upgrade to newer OS versions via
eos-updater-flatpak-autoinstall.d files in the OS.

Previously, Rhythmbox was part of the OSTree. In eos4.0 it was removed
from the OS, so an autoinstall file was added to make systems upgrading
to eos4.0 install the Flatpak to avoid the user losing a
previously-installed app.
https://github.com/endlessm/eos-application-tools/commit/0e6c484bb18e9b6360e0a51df114f00665bacb18

To avoid accidentally leaving OS components out of an image,
hooks/image/60-flatpak-autoinstall-counters.chroot asserts that any apps
listed as to be installed by autoinstall files are shipped as part of
the image. Such apps are listed in the 'apps_add_mandatory' key
separately from the 'apps_add_default' list so that personalities can
easily remove the default set without removing the mandatory set.

There is a checkpoint between eos4.0 and eos5, so it is guaranteed that
computers upgrading to eos5 will have first upgraded to eos4 (and hence
had Rhythmbox automatically installed if needed). So the autoinstall
file has been removed https://github.com/endlessm/eos-application-tools/commit/86436b4ea078f516c5a8bba529c1d6e387c6b401
and thus there is no need for Rhythmbox to be listed as mandatory any
more. We will continue to install it by default, except on base where we
will save a little bit of space.

https://phabricator.endlessm.com/T34075