kolibri: Retry importcontent on error
We see sporadic failures to import random pieces of content, due to the
Kolibri server returning something with the wrong checksum. (This is
believed to be the CDN serving an error, possibly a rate-limiting
error.) Typically, if we rerun the build, the error does not recur for
the same content file.

It is sad to fail a multi-hour, multi-hundred-gigabyte image build due
to a single broken content item. Retry `kolibri manage importcontent`
once, after a short pause, in hope of recovering in this case.

https://phabricator.endlessm.com/T34149
(cherry picked from commit 6eb5b59f89612006d8634e1cb3687caca47cb0cb)