config: Set eos-runtimes remote to use new repo URL
The legacy runtimes have been copied out of the eos repo to a new
eos-runtimes repo so that it can be treated as a pure flatpak repo. This
starts a long process where eventually we can remove the legacy runtimes
from the eos repo and treat it as a pure ostree repo.

https://phabricator.endlessm.com/T33445