buildscript: Declare repo as a git safe directory
Newer git prevents operations when the calling user is not the same as
the user owning the repo. Currently the only git operation run is `git
rev-parse HEAD` to record the builder commit in the manifest, but even
that is enough to make git fail. Add the repo as a safe directory inside
the build root since the builder is running as root and the repo may not
be owned by root.

https://community.endlessos.com/t/custom-build-problem-with-5-0/19900