content: Add hook to remotely fetch Kolibri content
If a custom Kolibri content server is in use and credentials are
available, use the API to import the desired channels remotely. When the
time comes to populate the channels on the image, all the content will
be available on the server.

https://phabricator.endlessm.com/T33705
(cherry picked from commit 03ffcc3840810f42ed278676aa29ade9a9a7afa3)