image: Ensure all available Kolibri content downloaded
By default, Kolibri only imports content nodes that the current instance
has a renderer for. Since we're operating on a homedir that may be
missing plugins that are available at runtime (such as the zim plugin),
it may skip nodes that we ultimately want at runtime. Use the
`--include-unrenderable-content` option so that no nodes are filtered
out.

Helps: https://github.com/endlessm/endless-key-content-private/issues/103