hooks: Request Kolibri content server fetch all nodes
By default, Kolibri filters the content nodes to download by the types
of content the instance can render. For the content server we want to
fully mirror the channel and download all nodes regardless of whether
the content can be rendered or not.

To do that, the API request is made with the `renderable_only` option
set to `False`. For better or worse, Kolibri 0.15 ignores unknown
options, so this will only do anything useful if the content server
handles the `renderable_only` option[1].

1. https://github.com/learningequality/kolibri/pull/11513

Helps: https://github.com/endlessm/endless-key-content-private/issues/103