Export AppStream catalogue for preinstalled Flatpak apps
Although we have added the name and summary for each app to the JSON
manifest, there is no substitute for having all the data.

It is better to capture this at build time rather than fetching it on
the fly later because apps can change or disappear, but we would like to
be able to describe exactly what is in any given image.

The Flatpak repos we use have already done the hard work of
validating every app and runtime's exported AppStream metainfo. So
we can use the local copy of each remote's AppStream to find the
<component> for each installed app and runtime, and concatenate these
into a new catalog. This also avoids any issues where our local version
of AppStream tools and libraries are more strict than those used on the
remotes.

XML compresses reasonably well. Running this script against the apps
installed on my development system, the uncompressed file is 7.0 MiB.
Compressing with pigz at the maximum compression level (as implemented
here) comes out at 2.0 MiB and takes 77ms. Compressing with zstd at its
default compression level (3) comes out at 1.9 MiB and takes 68ms.
Compressing with zstd at its maximum compression level (19 if you don't
pass --ultra which comes with some caveats) reduces the file to 1.4 MiB
but takes 3.7s. Honestly I was hoping that zstd would perform better so
I would have an excuse to start introducing it into eos-image-builder
but it's just not worth it for these minimal gains; plus, the
[AppStream Catalog XML specification][0] specifies GZip compression, so
there's that.

[0]: https://www.freedesktop.org/software/appstream/docs/chap-CatalogData.html

https://phabricator.endlessm.com/T35013