lib: Drop keyring creation
The only use of the GPG keys is to validate the buildroot source. With
`mmdebstrap` a full GPG keyring was required. However, the `ostree`
`--gpg-import` option accepts individual keys and builds the remote
trusted keyring on its own. Stop creating the keyring and just pass the
key paths to `ostree`. After that the explicit host `gnupg` dependency
is no longer required (although `ostree` internally uses `gnupg`).

https://phabricator.endlessm.com/T35019