fbe: Adopt upstream vendor.conf path
Previously we built GNOME Initial Setup to use the path
/var/lib/eos-image-defaults/branding/gnome-initial-setup.conf and
installed the keyfile specified in the '[image] branding_fbe_config' key
there. (There was no validation that its name was actually
gnome-initial-setup.conf, which was necessary for it to work.)

When we contributed this feature upstream to gnome-initial-setup, the
path was made configurable with a build-time parameter, with no default.

Some time later, Initial Setup started reading
/etc/gnome-initial-setup/vendor.conf if present; and
/usr/share/gnome-initial-setup/vendor.conf otherwise. It retained a
build-time parameter, which we continued to set; but setting that
parameter disabled *both* of the upstream paths.

With the upstream behaviour, it is possible to ship overrides in the
immutable bit of the OS, and further override them in /etc. This is in
my opinion preferable to what we have historically done, to
always write a file in /var for all images, even if the specific image
does not want any image-specific customisation of Initial Setup.

Install any config file to the upstream path in /etc, matching a build
configuration change I will make in our gnome-initial-setup package.

https://phabricator.endlessm.com/T35040