eibkolibri: Handle updated channels
When a channel is updated is updated on Kolibri Studio, importing the
channel and content again does not fetch new content nodes. Instead, a
channel update process is needed, which also requires generating the
diff between the old and new versions of the channel.

Fortunately, there are APIs for those tasks, so run them if the channel
already exists on the server to ensure the it has seeded all content
nodes from the (potentially) updated channel.

https://phabricator.endlessm.com/T34697