eibkolibri: Remove timeout option from tasks options
The tasks in neither Kolibri 0.15 nor 0.16 marshal the `timeout` option.
We don't actually need this, though, as the default timeout of 60
seconds is fine. If the server hasn't responded in 60 seconds, there's
probably something wrong. The long timeout was added back in 32ed0f2
when we were using Fastly as a Kolibri CDN and I believed it wouldn't
respond for a long time in our unusual configuration.

https://phabricator.endlessm.com/T34697