eibkolibri: Follow channel update with channel import
If new nodes are added in a new channel version, the channel needs to be
updated or they will be skipped if they're below an already existing
node. However, if there are missing nodes from a previously failed
import, then a regular import is needed. Therefore, follow a channel
update with a channel import.

The update is run first because the algorithm for calculating newly
needed nodes depends on comparing the old and new versions of the
channel metadata. One a regular import is done, the old version of the
channel metadata is replaced and the calculation can't be made anymore.

https://phabricator.endlessm.com/T34697