Initialise /var/lib and friends in buildroot
We recently merged a change to completely wipe `/var` in the ostree,
replacing it with an empty directory. This is because on a real system,
`/var` from the ostree is not used: instead a persistent, read-write
directory is mounted over it.

It turns out that `apt update` is happy to create its own directory
hierarchy below `/var/lib`, but if that directory does not exist it fails
with:

    E: List directory /var/lib/apt/lists/partial is missing. - Acquire (2: No such file or directory)

Similarly, installing packages requires /var/lib/dpkg to exist, but we
have moved this to /usr

    E: Could not open lock file /var/lib/dpkg/lock-frontend - open (2: No such file or directory)
    E: Unable to acquire the dpkg frontend lock (/var/lib/dpkg/lock-frontend), are you root?

On a real Endless OS system, `/var/lib` is initialised at boot by
`systemd-tmpfiles`. Do the same here. Move the dpkg database back to its
rightful home. Remove the tmpfiles snippet that eos-ostree-builder adds
to deal with the changes we are undoing.

https://phabricator.endlessm.com/T4482