Remove Cheese from mandatory apps
Cheese was moved from the OS to a Flatpak in eos4.0. There has been a
checkpoint between eos4.0 and eos5.0+, so we can now remove the
flatpak-autoinstall.d rule that requires it to be installed. It follows
that we can relegate it to "default" status, which in particular means
that base images need no longer include it.

https://phabricator.endlessm.com/T35175