Stop preinstalling GNOME Logs
In eos4.0 (and earlier), gnome-logs was baked into the OSTree. It is in
apps_add_mandatory because we used flatpak-autoinstall.d to cause the
Flatpak version to be installed when upgrading from eos4.0 to eos5.[01]
so that the app does not disappear with the upgrade.

However this is really quite a niche app, so we will just stop doing
that. Our metrics show that this app is very seldom used. People who
need to read the journal, and prefer this app to journalctl, can install
it from the App Center.

https://phabricator.endlessm.com/T35176