Remove desktop branding hook, aka watermark support
We are removing eos-watermark-extension. It is very lightly used, not
enough to justify updating it for GNOME 43 compatibility.

If we wish to bring this functionality back in future, we should instead
package https://pagure.io/background-logo-extension, which is actually
the ancestral source of the eos-watermark-extension code and which is
actively maintained by a core Shell developer for each Fedora release.

In that case, we would use GSettings overrides rather than this
gnome-shell.conf INI file to set the logo path.

https://phabricator.endlessm.com/T35077