image: Add ability to create QEMU qcow2 images
The QEMU qcow2 format[1][2] provides a lot of nice features compared to
raw disk images. Some of the most important ones:

* It's a copy-on-write format, so the file grows as data is added.
* It supports compression.
* It supports snapshots.

The first 2 mean that a minimized download can be made with a lot of
space to expand the root partition into. Probably most important to us,
this is the format used by GNOME Boxes (really QEMU via libvirt). That
means you can hand the image directly to boxes without creating a qcow2
disk image that the ISO installs to or converting a raw image to qcow2.

1. https://www.qemu.org/docs/master/system/images.html#cmdoption-image-formats-arg-qcow2
2. https://en.wikipedia.org/wiki/Qcow

https://phabricator.endlessm.com/T35332