config: Build qcow2 images for eos base
In my ideal world, we'd do this for all eos products and add them to
osinfo-db. That way we could streamline the GNOME Boxes experience where
it's just going to write all the data from the ISO to a qcow2 image
anyways.

However, doing this for the bigger personalities would add a lot of
overhead. We'd really want to choose one of OVF or qcow2 as "the VM
image" since doing both would be painful. For now, punt on that question
and just enable it for the small base image where it can essentially be
a power user tool.

https://phabricator.endlessm.com/T35332