Switch to preinstalling Arduino IDE v2
This is a total rewrite of the app, using Electron. Version 1 of the
IDE, which we were previously preinstalling, is no longer supported.

Sadly, the rewrite is not published for arm64. But we have always
incorrectly omitted version 1 of the IDE from arm64 (see previous
commit) so no great loss.

https://phabricator.endlessm.com/T35401