image: Chown Kolibri and Endless Key data
tmpfiles.d configuration is provided for both that changes ownership to
kolibri:kolibri for both at runtime. However, if there's an overlayfs on
/var like there is when booting an ISO, all the files are copied to the
upper tmpfs. That's likely to OOM the system when either is filled with
preloaded data.

One possibility is to run the tmpfiles.d configuration here. For
simplicity, we just run `chown` here. For Kolibri, a separate hook is
added since the content generation is split across 3 hooks.

https://phabricator.endlessm.com/T35498