Make Endless Key collections, Kolibri channels & node IDs merged fields
Now that we have inheritance in personality configs, it becomes useful
to be able to say that (for example) es_MX should add 1 specific channel
to the selection inherited from es.

This does have the side-effect of sorting the lists. I believe that for
channel IDs this will affect the order that the channels are shown in
the Kolibri Learn tab: rather than being in the order specified in a
single place in the config, they will be in an arbitrary order
determined by how their unique IDs happen to sort.

https://phabricator.endlessm.com/T34731