Make Endless Key collections, Kolibri channels & node IDs merged fields
Now that we have inheritance in personality configs, it becomes useful
to be able to say that (for example) es_MX should add 1 specific channel
to the selection inherited from es.

Since commit 59980bce, values in merged fields keep the order in which
they were first encountered. As a result, if (for example) the es
personality specifies channels B, D, C (in that order) and es_MX adds
channel A, then the resulting order is B, D, C, A. The order is
significant: it is the order that channels are shown in the Kolibri
Learn tab. So in this case it is not possible to move channel A to the
top of the list without overriding the whole list of channels (which is
possible but loses the benefit of this change). In practice, I only
found two cases where this happens; in one case I hardcoded the previous
order, and in the other case I just accepted the slight change.

https://phabricator.endlessm.com/T34731