ImageConfigParser: Don't sort merged values
The order of configuration values may be significant, so the parser
should not sort them. To do that, use a single counter for add and del
variants and emit the values that have positive counts without sorting.
This relies on the Python dict stable order, which has been the case
since Python 3.6.