Match selected Endless Key content collections correctly
EIB_ENDLESSKEY_COLLECTIONS is a whitespace (typically space or
newline)-separated list of collection names. These are matched against
collection filenames to determine what to fetch.

We have collection files named `extras-preload-0001.json` and
`spanish-extras-preload-0001.json`. The intention is that if `[endlesskey]
collections` contains `spanish-extras-preload` but not `extras-preload`, only
the latter will be imported, not the former.

However, the logic that matched filenames against the configured array did not
consider the case when one collection name is a substring of another. It would
incorrectly match `extras-preload-0001.json` against `spanish-extras-preload`
because the latter contains `extras-preload` as a substring.

Make the check more precise by checking that the desired collection name is
present in EIB_ENDLESSKEY_COLLECTIONS, surrounded by whitespace. Surround
EIB_ENDLESSKEY_COLLECTIONS itself with whitespace to make this check work for
the first or last collection in the list.

(The root cause of this bug is that shell scripts are not a suitable language
for non-trivial logic. I tried to use Bash arrays but apparently Bash does not
have a "contains" operator for arrays.)

https://phabricator.endlessm.com/T35668#1004203